Added Quit method to Polar and replaced event fire in HandleSDL.

+++ include/TextAsset.h
@@ -10,7 +10,7 @@ public:
 	static std::string Type() { return "text"; }
 	TextAsset(const uint32_t length, const std::string &text) : Asset("text"), length(length), text(text) {}
 	static TextAsset Load(const std::string &data) {
-		const uint32_t dataLength = data.length();
+		const uint32_t dataLength = static_cast<uint32_t>(data.length());
 		if(dataLength < 4) { ENGINE_THROW("missing data length"); }
 		const uint32_t length = swapbe(*reinterpret_cast<const uint32_t *>(data.c_str()));
 		if(length > dataLength - 4) { ENGINE_THROW("invalid data length"); }
